requirements.txt
python-dotenv
langchain-openai
crewai
streamlit
